+++ CHANGES.md
@@ -1,6 +1,6 @@
 # Unreleased
 
-- Added `union` and `intersection`.
+- Added `union` and `inter`.
 
 # v0.1.0 2019-03-07